requirements.txt
openai==1.100.2
pandas==2.3.1
openpyxl==3.1.5
python-dotenv==1.1.1